Keep all of the dummy's request-related tests in a separate file
That's cleaner and more concise.
